requirements.txt
Flask>=3.0.0
Flask-MySQL>=1.5.0
Flask-CORS>=3.1.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
